Abstracted leaderboard a bit
