+++ frontend/src/components/problemOccurenceOverview/contents/ProblemOccurrenceLeaderboard.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Container, Row, Col } from "reactstrap";
-import Leaderboard from "../../leaderboards/Leaderboard";
+import AggregateLeaderboard from "../../leaderboards/AggregateLeaderboard";
 
 /**
  * A component that includes the leaderboard of a problem occurence.
@@ -10,7 +10,7 @@ function ProblemOccurrenceLeaderboard({problemData, leaderboardData}) {
     <Container fluid className="ps-0 pt-2">
         <Row className="">
           <Col>
-            <Leaderboard problemData={problemData} leaderboardData={leaderboardData}/>
+            <AggregateLeaderboard problemData={problemData} leaderboardData={leaderboardData}/>
           </Col>
         </Row>
     </Container>
